geomfn: consolidate affine transformations
* Introduce `applyOnGeomT`, which applies a function on a set
  of coordinates.
* Consolidate Translate, Scale and Rotate in terms of Affine.
  This is what PostGIS does, and removes a lot of duplicate code.
* Implement ST_Affine.
* Add unimplemented errors for remaining ST_Rotate items.

Release justification: low risk updates to to new functionality.

Release note (sql change): Implement ST_Affine.
