+++ pkg/geo/geomfn/affine_transforms.go
@@ -1,0 +1,228 @@
+// Copyright 2020 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package geomfn
+
+import (
+	"math"
+
+	"github.com/cockroachdb/cockroach/pkg/geo"
+	"github.com/cockroachdb/errors"
+	"github.com/twpayne/go-geom"
+)
+
+// AffineMatrix defines an affine transformation matrix for a geom object.
+// It is expected to be of the form:
+//     a  b  c  x_off
+//     d  e  f  y_off
+//     g  h  i  z_off
+//     0  0  0  1
+// Which gets applies onto a coordinate of form:
+//     (x y z 0)^T
+// With the following transformation:
+//   x' = a*x + b*y + c*z + x_off
+//   y' = d*x + e*y + f*z + y_off
+//   z' = g*x + h*y + i*z + z_off
+type AffineMatrix [][]float64
+
+// Affine applies a 3D affine transformation onto the given geometry.
+// See: https://en.wikipedia.org/wiki/Affine_transformation.
+func Affine(g geo.Geometry, m AffineMatrix) (geo.Geometry, error) {
+	if g.Empty() {
+		return g, nil
+	}
+
+	t, err := g.AsGeomT()
+	if err != nil {
+		return geo.Geometry{}, err
+	}
+
+	newT, err := affine(t, m)
+	if err != nil {
+		return geo.Geometry{}, err
+	}
+	return geo.MakeGeometryFromGeomT(newT)
+}
+
+func affine(t geom.T, m AffineMatrix) (geom.T, error) {
+	return applyOnCoordsForGeomT(t, func(l geom.Layout, dst, src []float64) error {
+		var z float64
+		if l.ZIndex() != -1 {
+			z = src[l.ZIndex()]
+		}
+		newX := m[0][0]*src[0] + m[0][1]*src[1] + m[0][2]*z + m[0][3]
+		newY := m[1][0]*src[0] + m[1][1]*src[1] + m[1][2]*z + m[1][3]
+		newZ := m[2][0]*src[0] + m[2][1]*src[1] + m[2][2]*z + m[2][3]
+
+		dst[0] = newX
+		dst[1] = newY
+		if l.ZIndex() != -1 {
+			dst[2] = newZ
+		}
+		if l.MIndex() != -1 {
+			dst[l.MIndex()] = src[l.MIndex()]
+		}
+		return nil
+	})
+}
+
+// Translate returns a modified Geometry whose coordinates are incremented
+// or decremented by the deltas.
+func Translate(g geo.Geometry, deltas []float64) (geo.Geometry, error) {
+	if g.Empty() {
+		return g, nil
+	}
+
+	t, err := g.AsGeomT()
+	if err != nil {
+		return geo.Geometry{}, err
+	}
+
+	newT, err := translate(t, deltas)
+	if err != nil {
+		return geo.Geometry{}, err
+	}
+	return geo.MakeGeometryFromGeomT(newT)
+}
+
+func translate(t geom.T, deltas []float64) (geom.T, error) {
+	if t.Layout().Stride() != len(deltas) {
+		err := geom.ErrStrideMismatch{
+			Got:  len(deltas),
+			Want: t.Layout().Stride(),
+		}
+		return nil, errors.Wrap(err, "translating coordinates")
+	}
+	var zOff float64
+	if t.Layout().ZIndex() != -1 {
+		zOff = deltas[t.Layout().ZIndex()]
+	}
+	return affine(
+		t,
+		AffineMatrix([][]float64{
+			{1, 0, 0, deltas[0]},
+			{0, 1, 0, deltas[1]},
+			{0, 0, 1, zOff},
+			{0, 0, 0, 1},
+		}),
+	)
+}
+
+// Scale returns a modified Geometry whose coordinates are multiplied by the factors.
+func Scale(g geo.Geometry, factors []float64) (geo.Geometry, error) {
+	var zFactor float64
+	if len(factors) > 2 {
+		zFactor = factors[2]
+	}
+	return Affine(
+		g,
+		AffineMatrix([][]float64{
+			{factors[0], 0, 0, 0},
+			{0, factors[1], 0, 0},
+			{0, 0, zFactor, 0},
+			{0, 0, 0, 1},
+		}),
+	)
+}
+
+// ScaleRelativeToOrigin returns a modified Geometry whose coordinates are multiplied by the factors relative to the origin
+func ScaleRelativeToOrigin(
+	g geo.Geometry, factor geo.Geometry, origin geo.Geometry,
+) (geo.Geometry, error) {
+	if g.Empty() {
+		return g, nil
+	}
+
+	t, err := g.AsGeomT()
+	if err != nil {
+		return geo.Geometry{}, err
+	}
+
+	factorG, err := factor.AsGeomT()
+	if err != nil {
+		return geo.Geometry{}, err
+	}
+
+	factorPointG, ok := factorG.(*geom.Point)
+	if !ok {
+		return geo.Geometry{}, errors.Newf("the scaling factor must be a Point")
+	}
+
+	originG, err := origin.AsGeomT()
+	if err != nil {
+		return geo.Geometry{}, err
+	}
+
+	originPointG, ok := originG.(*geom.Point)
+	if !ok {
+		return geo.Geometry{}, errors.Newf("the false origin must be a Point")
+	}
+
+	if factorG.Stride() != originG.Stride() {
+		err := geom.ErrStrideMismatch{
+			Got:  factorG.Stride(),
+			Want: originG.Stride(),
+		}
+		return geo.Geometry{}, errors.Wrap(err, "number of dimensions for the scaling factor and origin must be equal")
+	}
+
+	// Offset by the origin, scale, and translate it back to the origin.
+	offsetDeltas := make([]float64, 0, 3)
+	offsetDeltas = append(offsetDeltas, -originPointG.X(), -originPointG.Y())
+	if originG.Layout().ZIndex() != -1 {
+		offsetDeltas = append(offsetDeltas, -originPointG.Z())
+	}
+	retT, err := translate(t, offsetDeltas)
+	if err != nil {
+		return geo.Geometry{}, err
+	}
+
+	xFactor, yFactor := factorPointG.X(), factorPointG.Y()
+	var zFactor float64 = 1
+	if factorPointG.Layout().ZIndex() != -1 {
+		zFactor = factorPointG.Z()
+	}
+	retT, err = affine(
+		retT,
+		AffineMatrix([][]float64{
+			{xFactor, 0, 0, 0},
+			{0, yFactor, 0, 0},
+			{0, 0, zFactor, 0},
+			{0, 0, 0, 1},
+		}),
+	)
+	if err != nil {
+		return geo.Geometry{}, err
+	}
+
+	for i := range offsetDeltas {
+		offsetDeltas[i] = -offsetDeltas[i]
+	}
+	retT, err = translate(retT, offsetDeltas)
+	if err != nil {
+		return geo.Geometry{}, err
+	}
+
+	return geo.MakeGeometryFromGeomT(retT)
+}
+
+// Rotate returns a modified Geometry whose coordinates are rotated
+// around the origin by a rotation angle.
+func Rotate(g geo.Geometry, rotRadians float64) (geo.Geometry, error) {
+	return Affine(
+		g,
+		AffineMatrix([][]float64{
+			{math.Cos(rotRadians), -math.Sin(rotRadians), 0, 0},
+			{math.Sin(rotRadians), math.Cos(rotRadians), 0, 0},
+			{0, 0, 1, 0},
+			{0, 0, 0, 1},
+		}),
+	)
+}
