+++ pkg/sql/logictest/testdata/logic_test/geospatial
@@ -4314,6 +4314,25 @@ POLYGON ((-0.1 0, 1 0, 1 1, -0.1 1, -0.1 0))           POLYGON ((-0.05 0, 0.5 0,
 POLYGON ((-1 0, 0 0, 0 1, -1 1, -1 0))                 POLYGON ((-0.5 0, 0 0, 0 2, -0.5 2, -0.5 0))           POLYGON ((-0.25 0, 0 0, 0 3, -0.25 3, -0.25 0))           POLYGON ((-3 -1, -1 -1, -1 1, -3 1, -3 -1))
 POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))                    POLYGON ((0 0, 0.5 0, 0.5 2, 0 2, 0 0))                POLYGON ((0 0, 0.25 0, 0.25 3, 0 3, 0 0))                 POLYGON ((-1 -1, 1 -1, 1 1, -1 1, -1 -1))
 
+query TT
+SELECT
+  ST_AsText(a.geom) d,
+  ST_AsText(ST_Affine(a.geom, 1, 2, 3, 4, 5, 6), 3)
+FROM geom_operators_test a
+ORDER BY d ASC
+----
+NULL                                                   NULL
+GEOMETRYCOLLECTION (GEOMETRYCOLLECTION (POINT (0 0)))  GEOMETRYCOLLECTION (GEOMETRYCOLLECTION (POINT (5 6)))
+GEOMETRYCOLLECTION EMPTY                               GEOMETRYCOLLECTION EMPTY
+LINESTRING (-0.5 0.5, 0.5 0.5)                         LINESTRING (5.5 6.5, 6.5 9.5)
+LINESTRING EMPTY                                       LINESTRING EMPTY
+POINT (-0.5 0.5)                                       POINT (5.5 6.5)
+POINT (0.5 0.5)                                        POINT (6.5 9.5)
+POINT (5 5)                                            POINT (20 41)
+POINT EMPTY                                            POINT EMPTY
+POLYGON ((-0.1 0, 1 0, 1 1, -0.1 1, -0.1 0))           POLYGON ((4.9 5.7, 6 9, 8 13, 6.9 9.7, 4.9 5.7))
+POLYGON ((-1 0, 0 0, 0 1, -1 1, -1 0))                 POLYGON ((4 3, 5 6, 7 10, 6 7, 4 3))
+POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))                    POLYGON ((5 6, 6 9, 8 13, 7 10, 5 6))
 
 query T
 SELECT ST_Summary('POINT(0 0)'::geometry)
@@ -4704,15 +4723,15 @@ SELECT
 FROM geom_operators_test a
 ORDER BY d ASC
 ----
-NULL                                                   NULL                                                    NULL
-GEOMETRYCOLLECTION (GEOMETRYCOLLECTION (POINT (0 0)))  GEOMETRYCOLLECTION (GEOMETRYCOLLECTION (POINT (-0 0)))  GEOMETRYCOLLECTION (GEOMETRYCOLLECTION (POINT (0 0)))
-GEOMETRYCOLLECTION EMPTY                               GEOMETRYCOLLECTION EMPTY                                GEOMETRYCOLLECTION EMPTY
-LINESTRING (-0.5 0.5, 0.5 0.5)                         LINESTRING (0.5 -0.5, -0.5 -0.5)                        LINESTRING (-0.707106781186547 0, 0 0.707106781186547)
-LINESTRING EMPTY                                       LINESTRING EMPTY                                        LINESTRING EMPTY
-POINT (-0.5 0.5)                                       POINT (0.5 -0.5)                                        POINT (-0.707106781186547 0)
-POINT (0.5 0.5)                                        POINT (-0.5 -0.5)                                       POINT (0 0.707106781186547)
-POINT (5 5)                                            POINT (-5.000000000000001 -4.999999999999999)           POINT (0 7.071067811865476)
-POINT EMPTY                                            POINT EMPTY                                             POINT EMPTY
-POLYGON ((-0.1 0, 1 0, 1 1, -0.1 1, -0.1 0))           POLYGON ((0.1 -0, -1 0, -1 -1, 0.1 -1, 0.1 -0))         POLYGON ((-0.070710678118655 -0.070710678118655, 0.707106781186548 0.707106781186547, 0 1.414213562373095, -0.777817459305202 0.636396103067893, -0.070710678118655 -0.070710678118655))
-POLYGON ((-1 0, 0 0, 0 1, -1 1, -1 0))                 POLYGON ((1 -0, -0 0, -0 -1, 1 -1, 1 -0))               POLYGON ((-0.707106781186548 -0.707106781186547, 0 0, -0.707106781186547 0.707106781186548, -1.414213562373095 0, -0.707106781186548 -0.707106781186547))
-POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))                    POLYGON ((-0 0, -1 0, -1 -1, -0 -1, -0 0))              POLYGON ((0 0, 0.707106781186548 0.707106781186547, 0 1.414213562373095, -0.707106781186547 0.707106781186548, 0 0))
+NULL                                                   NULL                                                   NULL
+GEOMETRYCOLLECTION (GEOMETRYCOLLECTION (POINT (0 0)))  GEOMETRYCOLLECTION (GEOMETRYCOLLECTION (POINT (0 0)))  GEOMETRYCOLLECTION (GEOMETRYCOLLECTION (POINT (0 0)))
+GEOMETRYCOLLECTION EMPTY                               GEOMETRYCOLLECTION EMPTY                               GEOMETRYCOLLECTION EMPTY
+LINESTRING (-0.5 0.5, 0.5 0.5)                         LINESTRING (0.5 -0.5, -0.5 -0.5)                       LINESTRING (-0.707106781186547 0, 0 0.707106781186547)
+LINESTRING EMPTY                                       LINESTRING EMPTY                                       LINESTRING EMPTY
+POINT (-0.5 0.5)                                       POINT (0.5 -0.5)                                       POINT (-0.707106781186547 0)
+POINT (0.5 0.5)                                        POINT (-0.5 -0.5)                                      POINT (0 0.707106781186547)
+POINT (5 5)                                            POINT (-5.000000000000001 -4.999999999999999)          POINT (0 7.071067811865476)
+POINT EMPTY                                            POINT EMPTY                                            POINT EMPTY
+POLYGON ((-0.1 0, 1 0, 1 1, -0.1 1, -0.1 0))           POLYGON ((0.1 -0, -1 0, -1 -1, 0.1 -1, 0.1 -0))        POLYGON ((-0.070710678118655 -0.070710678118655, 0.707106781186548 0.707106781186547, 0 1.414213562373095, -0.777817459305202 0.636396103067893, -0.070710678118655 -0.070710678118655))
+POLYGON ((-1 0, 0 0, 0 1, -1 1, -1 0))                 POLYGON ((1 -0, 0 0, -0 -1, 1 -1, 1 -0))               POLYGON ((-0.707106781186548 -0.707106781186547, 0 0, -0.707106781186547 0.707106781186548, -1.414213562373095 0, -0.707106781186548 -0.707106781186547))
+POLYGON ((0 0, 1 0, 1 1, 0 1, 0 0))                    POLYGON ((0 0, -1 0, -1 -1, -0 -1, 0 0))               POLYGON ((0 0, 0.707106781186548 0.707106781186547, 0 1.414213562373095, -0.707106781186547 0.707106781186548, 0 0))
